add apply method for adding seasonal stats to legacy playlist containers

+++ DragonFruit.Six.Api/Legacy/Entities/LegacyStats.cs
@@ -37,91 +37,90 @@ namespace DragonFruit.Six.Api.Legacy.Entities
         /// Total barricades built
         /// </summary>
         [JsonProperty("barricades")]
-        public uint Barricades { get; set; }
+        public int Barricades { get; set; }
 
         /// <summary>
         /// Total reinforcements deployed
         /// </summary>
         [JsonProperty("reinforcements")]
-        public uint Reinforcements { get; set; }
+        public int Reinforcements { get; set; }
 
         /// <summary>
         /// Total gadgets destroyed
         /// </summary>
         [JsonProperty("gadgets_destroyed")]
-        public uint GadgetsDestroyed { get; set; }
+        public int GadgetsDestroyed { get; set; }
 
         /// <summary>
         /// Number of opponents downed
         /// </summary>
         [JsonProperty("downs")]
-        public uint Downs { get; set; }
+        public int Downs { get; set; }
 
         /// <summary>
         /// Number of teammates revived from a downed state
         /// </summary>
         [JsonProperty("revives")]
-        public uint Revives { get; set; }
+        public int Revives { get; set; }
 
         /// <summary>
         /// Number of times the player has killed themself (C4, jumping off a building, etc.)
         /// </summary>
         [JsonProperty("suicides")]
-        public uint Suicides { get; set; }
+        public int Suicides { get; set; }
 
         /// <summary>
         /// Total headshot kills
         /// </summary>
         [JsonProperty("headshots")]
-        public uint Headshots { get; set; }
+        public int Headshots { get; set; }
 
         /// <summary>
         /// Total wallbangs
         /// </summary>
         [JsonProperty("penetrations")]
-        public uint Penetrations { get; set; }
+        public int Penetrations { get; set; }
 
         /// <summary>
         /// Total kills when unable to see (i.e. a flashbang or an angry blitz)
         /// </summary>
         [JsonProperty("blind_kills")]
-        public uint BlindKills { get; set; }
+        public int BlindKills { get; set; }
 
         /// <summary>
         /// Total knife kills
         /// </summary>
         [JsonProperty("melee")]
-        public uint Knifes { get; set; }
+        public int Knifes { get; set; }
 
         /// <summary>
         /// Total kill assists
         /// </summary>
         [JsonProperty("assists")]
-        public uint Assists { get; set; }
+        public int Assists { get; set; }
 
         /// <summary>
         /// Total down assists (assisting in downing an opponent)
         /// </summary>
-        // todo convert to ulong
         [JsonProperty("down_assists")]
-        public uint DownAssists { get; set; }
+        public long DownAssists { get; set; }
 
         /// <summary>
         /// Total shots fired
         /// </summary>
         [JsonProperty("shots_fired")]
-        public ulong ShotsFired { get; set; }
+        public long ShotsFired { get; set; }
 
         /// <summary>
         /// Total shots connected
         /// </summary>
         [JsonProperty("shots_connected")]
-        public ulong ShotsConnected { get; set; }
+        public long ShotsConnected { get; set; }
 
         /// <summary>
         /// Total points gained from all modes
         /// </summary>
         [JsonProperty("experience")]
-        public ulong Experience { get; set; }
+        public long Experience { get; set; }
     }
 }
